+++ nimbus/examples/experiment.rs
@@ -64,8 +64,16 @@ fn main() -> Result<()> {
                 .about("Show all experiments, followed by the enrolled experiments"),
         )
         .subcommand(
+            SubCommand::with_name("fetch-experiments")
+            .about("Fetch experiments from the server. Subsequent calls to apply-pending-experiments will change enrolments."),
+        )
+        .subcommand(
+            SubCommand::with_name("apply-pending-experiments")
+            .about("Updates enrollments with the experiments last fetched from the server with fetch-experiments"),
+        )
+        .subcommand(
             SubCommand::with_name("update-experiments")
-            .about("Updates experiments and enrollments from the server"),
+            .about("Equivalent to fetch-experiments and apply-pending-experiments together"),
         )
         .subcommand(
             SubCommand::with_name("opt-in")
@@ -199,7 +207,8 @@ fn main() -> Result<()> {
     let mut nimbus_client = NimbusClient::new(context, "", Some(config), aru)?;
 
     // Explicitly update experiments at least once for init purposes
-    nimbus_client.update_experiments()?;
+    nimbus_client.fetch_experiments()?;
+    nimbus_client.apply_pending_experiments()?;
 
     // We match against the subcommands
     match matches.subcommand() {
@@ -223,10 +232,21 @@ fn main() -> Result<()> {
                     )
                 });
         }
+        ("fetch-experiments", _) => {
+            println!("======================================");
+            println!("Fetching experiments");
+            nimbus_client.fetch_experiments()?;
+        }
+        ("apply-pending-experiments", _) => {
+            println!("======================================");
+            println!("Applying pending experiments");
+            nimbus_client.apply_pending_experiments()?;
+        }
         ("update-experiments", _) => {
             println!("======================================");
-            println!("Updating experiments");
-            nimbus_client.update_experiments()?;
+            println!("Fetching and applying experiments");
+            nimbus_client.fetch_experiments()?;
+            nimbus_client.apply_pending_experiments()?;
         }
         ("opt-in", Some(matches)) => {
             println!("======================================");