+++ nimbus/src/client/http_client.rs
@@ -109,56 +109,62 @@ impl SettingsClient for Client {
         unimplemented!();
     }
 
-    fn get_experiments(&mut self) -> Result<Vec<Experiment>> {
+    fn fetch_experiments(&mut self) -> Result<Vec<Experiment>> {
         let path = format!(
             "buckets/{}/collections/{}/records",
             &self.bucket_name, &self.collection_name
         );
         let url = self.base_url.join(&path)?;
         let req = Request::get(url);
-        // We first encode the response into a `serde_json::Value`
-        // to allow us to deserialize each experiment individually,
-        // omitting any malformed experiments
-        let resp = self.make_request(req)?.json::<serde_json::Value>()?;
-        let data = resp.get("data").ok_or(Error::InvalidExperimentResponse)?;
-        let mut res = Vec::new();
-        for exp in data.as_array().ok_or(Error::InvalidExperimentResponse)? {
-            // Validate the schema major version matches the supported version
-            let exp_schema_version = match exp.get("schemaVersion") {
-                Some(ver) => serde_json::from_value::<String>(ver.to_owned())
-                    .unwrap_or_else(|_| "".to_string()),
-                None => {
-                    log::trace!("Missing schemaVersion: {:#?}", exp);
-                    continue;
-                }
-            };
-            let schema_maj_version = exp_schema_version.split('.').next().unwrap_or("");
-            // While "0" is a valid schema version, we have already passed that so reserving zero as
-            // a special value here in order to avoid a panic, and just ignore the experiment.
-            let schema_version: u32 = schema_maj_version.parse().unwrap_or(0);
-            if schema_version != SCHEMA_VERSION {
-                log::info!(
+        let resp = self.make_request(req)?;
+        parse_experiments(&resp.text())
+    }
+}
+
+pub fn parse_experiments(payload: &str) -> Result<Vec<Experiment>> {
+    // We first encode the response into a `serde_json::Value`
+    // to allow us to deserialize each experiment individually,
+    // omitting any malformed experiments
+    let value: serde_json::Value = serde_json::from_str(payload)?;
+    let data = value.get("data").ok_or(Error::InvalidExperimentFormat)?;
+    let mut res = Vec::new();
+    for exp in data.as_array().ok_or(Error::InvalidExperimentFormat)? {
+        // Validate the schema major version matches the supported version
+        let exp_schema_version = match exp.get("schemaVersion") {
+            Some(ver) => {
+                serde_json::from_value::<String>(ver.to_owned()).unwrap_or_else(|_| "".to_string())
+            }
+            None => {
+                log::trace!("Missing schemaVersion: {:#?}", exp);
+                continue;
+            }
+        };
+        let schema_maj_version = exp_schema_version.split('.').next().unwrap_or("");
+        // While "0" is a valid schema version, we have already passed that so reserving zero as
+        // a special value here in order to avoid a panic, and just ignore the experiment.
+        let schema_version: u32 = schema_maj_version.parse().unwrap_or(0);
+        if schema_version != SCHEMA_VERSION {
+            log::info!(
                     "Schema version mismatch: Expected version {}, discarding experiment with version {}",
                     SCHEMA_VERSION, schema_version
                 );
-                // Schema version mismatch
-                continue;
-            }
+            // Schema version mismatch
+            continue;
+        }
 
-            match serde_json::from_value::<Experiment>(exp.clone()) {
-                Ok(exp) => res.push(exp),
-                Err(e) => {
-                    log::trace!("Malformed experiment data: {:#?}", exp);
-                    log::warn!(
-                        "Malformed experiment found! Experiment {},  Error: {}",
-                        exp.get("id").unwrap_or(&serde_json::json!("ID_NOT_FOUND")),
-                        e
-                    );
-                }
+        match serde_json::from_value::<Experiment>(exp.clone()) {
+            Ok(exp) => res.push(exp),
+            Err(e) => {
+                log::trace!("Malformed experiment data: {:#?}", exp);
+                log::warn!(
+                    "Malformed experiment found! Experiment {},  Error: {}",
+                    exp.get("id").unwrap_or(&serde_json::json!("ID_NOT_FOUND")),
+                    e
+                );
             }
         }
-        Ok(res)
     }
+    Ok(res)
 }
 
 #[cfg(test)]
@@ -238,7 +244,7 @@ mod tests {
     }
 
     #[test]
-    fn test_get_experiments_from_schema() {
+    fn test_fetch_experiments_from_schema() {
         viaduct_reqwest::use_reqwest_backend();
         // There are two experiments defined here, one has a "newer" schema version
         // in order to test filtering of unsupported schema versions.
@@ -256,7 +262,8 @@ mod tests {
             collection_name: "messaging-experiments".to_string(),
         };
         let mut http_client = Client::new(config).unwrap();
-        let resp = http_client.get_experiments().unwrap();
+        let resp = http_client.fetch_experiments().unwrap();
+
         m.expect(1).assert();
         assert_eq!(resp.len(), 1);
         let exp = &resp[0];
@@ -317,8 +324,8 @@ mod tests {
             collection_name: "messaging-experiments".to_string(),
         };
         let mut http_client = Client::new(config).unwrap();
-        assert!(http_client.get_experiments().is_ok());
-        let second_request = http_client.get_experiments();
+        assert!(http_client.fetch_experiments().is_ok());
+        let second_request = http_client.fetch_experiments();
         assert!(matches!(second_request, Err(Error::BackoffError(_))));
         m.expect(1).assert();
     }
@@ -340,8 +347,8 @@ mod tests {
             collection_name: "messaging-experiments".to_string(),
         };
         let mut http_client = Client::new(config).unwrap();
-        assert!(http_client.get_experiments().is_err());
-        let second_request = http_client.get_experiments();
+        assert!(http_client.fetch_experiments().is_err());
+        let second_request = http_client.fetch_experiments();
         assert!(matches!(second_request, Err(Error::BackoffError(_))));
         m.expect(1).assert();
     }
@@ -369,7 +376,7 @@ mod tests {
             duration: Duration::from_secs(30),
         };
         assert!(matches!(
-            http_client.get_experiments(),
+            http_client.fetch_experiments(),
             Err(Error::BackoffError(_))
         ));
         // Then do the actual test.
@@ -377,7 +384,7 @@ mod tests {
             observed_at: Instant::now() - Duration::from_secs(31),
             duration: Duration::from_secs(30),
         };
-        assert!(http_client.get_experiments().is_ok());
+        assert!(http_client.fetch_experiments().is_ok());
         m.expect(1).assert();
     }
 }