+++ nimbus/src/client/null_client.rs
@@ -20,8 +20,8 @@ impl SettingsClient for NullClient {
     fn get_experiments_metadata(&self) -> Result<String> {
         unimplemented!();
     }
-    fn get_experiments(&mut self) -> Result<Vec<Experiment>> {
-        Ok(vec![])
+    fn fetch_experiments(&mut self) -> Result<Vec<Experiment>> {
+        Ok(Default::default())
     }
 }
 
@@ -36,7 +36,9 @@ fn test_null_client() -> Result<()> {
     let tmp_dir = TempDir::new("test_null_client-test_null")?;
 
     let aru = Default::default();
-    let client = NimbusClient::new(Default::default(), tmp_dir.path(), None, aru)?;
+    let mut client = NimbusClient::new(Default::default(), tmp_dir.path(), None, aru)?;
+    client.fetch_experiments()?;
+    client.apply_pending_experiments()?;
 
     let experiments = client.get_all_experiments()?;
     assert_eq!(experiments.len(), 0);