+++ nimbus/src/persistence.rs
@@ -58,6 +58,7 @@ pub enum StoreId {
     Experiments,
     Enrollments,
     Meta,
+    Updates,
 }
 
 /// A wrapper for an Rkv store. Implemented to allow any value which supports
@@ -140,6 +141,7 @@ pub struct Database {
     meta_store: SingleStore,
     experiment_store: SingleStore,
     enrollment_store: SingleStore,
+    updates_store: SingleStore,
 }
 
 impl Database {
@@ -152,11 +154,13 @@ impl Database {
         let meta_store = rkv.open_single("meta", StoreOptions::create())?;
         let experiment_store = rkv.open_single("experiments", StoreOptions::create())?;
         let enrollment_store = rkv.open_single("enrollments", StoreOptions::create())?;
+        let updates_store = rkv.open_single("updates", StoreOptions::create())?;
         let db = Self {
             rkv,
             meta_store: SingleStore::new(meta_store),
             experiment_store: SingleStore::new(experiment_store),
             enrollment_store: SingleStore::new(enrollment_store),
+            updates_store: SingleStore::new(updates_store),
         };
         db.maybe_upgrade()?;
         Ok(db)
@@ -182,6 +186,11 @@ impl Database {
                 self.enrollment_store.clear(&mut writer)?;
             }
         }
+        // It is safe to clear the update store (i.e. the pending experiments) on all schema upgrades
+        // as it will be re-filled from the server on the next `fetch_experiments()`.
+        // The current contents of the update store may cause experiments to not load, or worse,
+        // accidentally unenrol.
+        self.updates_store.clear(&mut writer)?;
         self.meta_store
             .put(&mut writer, DB_KEY_DB_VERSION, &DB_VERSION)?;
         writer.commit()?;
@@ -195,6 +204,7 @@ impl Database {
             StoreId::Meta => &self.meta_store,
             StoreId::Experiments => &self.experiment_store,
             StoreId::Enrollments => &self.enrollment_store,
+            StoreId::Updates => &self.updates_store,
         }
     }