+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Unreleased changes
+## What's New
+ - Split up `NimbusClient.update_experiments()` into a slow `NimbusClient.fetch_experiments()` and a fast `NimbusClient.apply_pending_experiments()` to help apps manage concurrency and mutable state.
+ - Add `set_local_experiments(string)`, to help apps, build tooling for tests, and help during startup on first time run.
 
+## ⚠️ Breaking changes ⚠️
+ - `NimbusClient.updateExperiments()` is removed.
+ - Renamed `InvalidExperimentResponse` error to `InvalidExperimentFormat`.
 # 0.6.4 (_2020-12-16_)
 
 ## What's New